Separate all tests into classes
