+++ kin-backup-and-restore/kin-backup-and-restore-ui-tests/build.gradle
@@ -17,7 +17,7 @@ android {
 
     buildTypes {
         debug {
-            testCoverageEnabled true
+            testCoverageEnabled false
         }
     }
